Refactor: 이미지 관련 예외 패키지 변경
